# com.consultec.esigns.listener/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: "Esta es la definicion del API Rest para el manejo de documentos firmados digital y electronicamente, desarrollado por Consultec para Banco General"
  version: "1.0.0"
  title: "Signature capture listener"
  contact:
    email: "hrodriguez@consultec-ti.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/"
tags:
- name: "files"
  description: "Gestion de documentos"
  externalDocs:
    description: "Todo lo relacionado con el manejo de documentos"
    url: "http://localhost:8080/files"
schemes:
- "http"
paths:
  /files/receive:
    post:
      tags:
      - "receive"
      summary: "Recibe una representacion de un documento PDF codificado en el esquema Base64"
      operationId: "receive"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Cuerpo del paquete en donde se especifican las distintas caracteristicas del documento"
        required: true
        schema:
          $ref: "#/definitions/PayloadTO"
      responses:
        405:
          description: "Invalid input"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
  /files/upload:
    post:
      tags:
      - "upload"
      summary: "carga un archivo PDF"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  PayloadTO:
    type: "object"
    required:
    - "sessionID"
    - "origin"
    properties:
      stage:
          type: "string"
          enum:
          - "INIT"
          - "MANUAL_SIGNED"
          - "E_SIGNED"
          - "COMPLETED"
          default: "INIT"
      plainDocEncoded:
        type: "string"
      strokedDocEncoded:
        type: "string"
      signedDocEncoded:
          type: "string"
      strokes:
          type: "array"
          items:
            type: "string"
      images:
          type: "array"
          items:
            type: "string"
      sessionID: 
          type: "string"
      origin:
          type: "string"
      userLogged:
          type: "string"
      cookieHeader:
          type: "string"
      code:
          type: "string"
    xml:
      name: "PayloadTO"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"